+++ src/module/applications/names/names/cossack.ts
