+++ src/module/applications/names/names/czech.ts
