+++ src/module/applications/names/names/french.ts
