+++ src/module/applications/names/names/german.ts
