+++ src/module/applications/names/names/gypsy.ts
