+++ src/module/applications/names/names/hungarian.ts
