+++ src/module/applications/names/names/polish.ts
