+++ src/module/applications/names/names/spanish.ts
