+++ src/module/applications/names/names/swedish.ts
