+++ src/module/applications/sheets/actor-sheet.ts
@@ -8,8 +8,8 @@ import { HVItem } from '../../documents/item';
 import { CharacterActorData } from '../../../types/actor-types';
 import { Utils } from '../../utils/utils';
 import { HVDice } from '../../dice';
-import { HVNameGenerator } from '../../apps/names';
-import { HVPDF } from '../../pdf';
+import { HVNameGenerator } from '../../applications/names';
+import { HVPDF } from '../pdf';
 import { NPCGenerator } from '../npcgen';
 import { slideToggle } from '../../utils/slide';
 import { ContainerSheet } from '../../applications/sheets/container-sheet';
