+++ src/module/helveczia.ts
@@ -49,12 +49,12 @@ import { HV } from './config';
 import { registerHandlebarHelpers } from './handlebar-helpers';
 import { HVChat } from './chat';
 import { HVCombat, HVCombatant } from './combat';
-import { HVCardsHand, HVCardsPile, HVCardsControl } from './apps/cards';
-import { HVNameGenerator } from './apps/names';
+import { HVCardsHand, HVCardsPile, HVCardsControl } from './applications/cards';
+import { HVNameGenerator } from './applications/names';
 import { Utils } from './utils/utils';
 import { init as quench_tests_init } from '../tests/quench';
 import { registerKeyBindings } from './keys';
-import { HVToken } from './token';
+import { HVToken } from './documents/token';
 import { HVSceneConfig } from './scene';
 import { FrameView } from './utils/frameview';
 import { HVItem } from './documents/item';
